Implementation of additional features, completed for this semester, now requiring styling and formatting
Created features are located in the MyStuff page, specifically the Questions for My Provider Checklist, the notes app, as well as removing unnessarcy code or lines for the Quiz app (i.e. isOption is never called and not required for the if statement). Will add comments for replication purposes and clarity for future teams, also to add documentation for the purpose of pushing